Reset branch to only include accessibility icons
